+++ packages/furo-ui5/demos/demos.js
@@ -30,6 +30,7 @@ import './demo-furo-ui5-data-toggle-button.js';
 import './demo-furo-ui5-data-property.js';
 import './demo-furo-ui5-message-strip-display.js';
 import './demo-furo-ui5-notification-list-display.js';
+import './demo-furo-ui5-notification-group-display.js';
 import './demo-furo-ui5-data-ro.js';
 import './demo-furo-fat-type.js';
 import './demo-furo-ui5-button.js';