+++ packages/furo-ui5/src/furo-ui5-data-number-input.js
@@ -114,12 +114,6 @@ export class FuroUi5DataNumberInput extends FieldNodeAdapter(Input.default) {
         this.value = 0;
       }
     });
-
-    // created to avoid the default messages from ui5
-    this._valueStateElement = document.createElement('div');
-    this._valueStateElement.setAttribute('slot', 'valueStateMessage');
-    // eslint-disable-next-line wc/no-constructor-attributes
-    this.appendChild(this._valueStateElement);
   }
 
   /**
@@ -131,6 +125,18 @@ export class FuroUi5DataNumberInput extends FieldNodeAdapter(Input.default) {
     // eslint-disable-next-line wc/guard-super-call
     super.connectedCallback();
     this.readAttributes();
+
+    // created to avoid the default messages from ui5
+    const vse = this.querySelector('div[slot="valueStateMessage"]');
+    if (vse === null) {
+      this._valueStateElement = document.createElement('div');
+      this._valueStateElement.setAttribute('slot', 'valueStateMessage');
+      // eslint-disable-next-line wc/no-constructor-attributes
+      this.appendChild(this._valueStateElement);
+    } else {
+      this._valueStateElement = vse;
+      this._previousValueState.message = vse.innerText;
+    }
   }
 
   // overwrite. fix for ui5 input error under rc14
@@ -151,9 +157,6 @@ export class FuroUi5DataNumberInput extends FieldNodeAdapter(Input.default) {
     this._previousValueState.state = this.getAttribute('value-state')
       ? this.getAttribute('value-state')
       : 'None';
-    this._previousValueState.message = this.getAttribute('value-state-message')
-      ? this.getAttribute('value-state-message')
-      : '';
     // save the original attribute for later usages, we do this, because some components reflect
     Object.keys(this._privilegedAttributes).forEach(attr => {
       this._privilegedAttributes[attr] = this.getAttribute(attr);
@@ -184,6 +187,10 @@ export class FuroUi5DataNumberInput extends FieldNodeAdapter(Input.default) {
         // remove empty state
         if (this._tmpFAT.labels && this._tmpFAT.labels.empty) {
           delete this._tmpFAT.labels.empty;
+        }
+        // init labels in_tmpFAT
+        if (this._tmpFAT.labels === null) {
+          this._tmpFAT.labels = {};
         }
         // set modified on changes
         this._tmpFAT.labels.modified = true;
@@ -289,7 +296,11 @@ export class FuroUi5DataNumberInput extends FieldNodeAdapter(Input.default) {
     // suggestions
     // see Properties/Attributes from ui5 on https://sap.github.io/ui5-webcomponents/playground/components/Input/
     if (fatAttributes.suggestions !== undefined) {
-      this._setSuggestions(JSON.parse(fatAttributes.suggestions));
+      if (typeof fatAttributes.suggestions === 'string') {
+        this._setSuggestions(JSON.parse(fatAttributes.suggestions));
+      } else if (Array.isArray(fatAttributes.suggestions)) {
+        this._setSuggestions(fatAttributes.suggestions);
+      }
     }
 
     // icon