+++ packages/furo-ui5/src/furo-ui5-message-strip.js
@@ -117,6 +117,19 @@ class FuroUi5MessageStrip extends FBP(LitElement) {
   }
 
   /**
+   * Compatibility function to avoid breaking change
+   * @deprecated
+   * @param msg
+   */
+  show(msg) {
+    this.showInformation(msg);
+    // eslint-disable-next-line no-console
+    console.warn(
+      'Deprecated. Instead, use the explicit show functions such as showInformation, showError, showSuccess or showWarning.',
+    );
+  }
+
+  /**
    * shows an information message
    * if the param msg is empty, the attribute message is used.
    * @param msg