+++ packages/furo-icon/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.35](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.34...@furo/icon@1.3.35) (2021-05-05)
+
+**Note:** Version bump only for package @furo/icon
+
+
+
+
+
 ## [1.3.34](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.33...@furo/icon@1.3.34) (2021-04-01)
 
 **Note:** Version bump only for package @furo/icon